feature: move helper functions in backend/github_client into a separate file
