+++ backend/github_client.py
@@ -3,7 +3,7 @@ import os
 import time
 import httpx
 import asyncio
-import logging
+import backend.helpers as helpers
 
 from urllib.parse import urlencode
 from backend.exceptions import GitHubAPIError, GitHubTimeOutError
@@ -18,34 +18,6 @@ from backend.exceptions import GitHubAPIError, GitHubTimeOutError
 GITHUB_TOKEN = os.getenv("GITHUB_TOKEN")
 
 semaphore = asyncio.Semaphore(4)
-
-
-def parse_data(data: list[dict] | dict | None) -> list[dict]:
-    """Converts the data received from the GitHub API into a list[dict]
-
-    Args:
-        data (list[dict] | dict | None): The data obtained from GitHub API.
-
-    Returns:
-        list[dict]: The data as a list of dictionaries.
-    """
-    if isinstance(data, list):
-        return data
-
-    if data is None:
-        return []
-
-    # deleting unwanted keys
-    del data["incomplete_results"]
-    del data["repository_selection"]
-    del data["total_count"]
-
-    # getting the first object of data
-    keys = iter(data)
-    first_key = next(keys)
-    data = data[first_key]
-
-    return data
 
 
 async def get_paginated_data(
@@ -88,7 +60,7 @@ async def get_paginated_data(
         if response.status_code != 200:
             raise GitHubAPIError()
 
-        parsed_data = parse_data(response.json())
+        parsed_data = helpers.parse_data(response.json())
         data += parsed_data
         if "link" not in response.headers:
             break
@@ -101,33 +73,6 @@ async def get_paginated_data(
             url = match.group(1)
 
     return data
-
-
-def get_owner_and_reponame(repo_url: str) -> tuple:
-    owner_and_repo = repo_url.removeprefix("https://github.com/")
-    owner, _, repo_name = owner_and_repo.partition("/")
-
-    return owner, repo_name
-
-
-def parse_url(repo_url: str, target: str = None):
-    """Parses a repository URL into the GitHub API format:
-        https://api.github.com/repos/{owner}/{repo}
-
-    Args:
-        repo_url (str): The URL in the format https://github.com/user/repo.
-        target (str): Additional term to specify which GitHub API endpoint to hit.
-
-    Returns:
-         str: a parsed URL ready to be used to make a call to GitHub's API.
-    """
-    owner, repo_name = get_owner_and_reponame(repo_url)
-    url = f"https://api.github.com/repos/{owner}/{repo_name}"
-
-    if target is not None:
-        url += "/" + target
-
-    return url
 
 
 async def get_commits(repo_url: str) -> list[dict]:
@@ -143,7 +88,7 @@ async def get_commits(repo_url: str) -> list[dict]:
         GitHubAPIError: If the request to GitHub failed.
     """
     commits = []
-    url = parse_url(repo_url, "commits")
+    url = helpers.parse_url(repo_url, "commits")
 
     commits = await get_paginated_data(url)
 
@@ -168,7 +113,7 @@ async def get_issues(
         GitHubAPIError: If the request to GitHub failed.
     """
     issues = []
-    url = parse_url(repo_url, "issues")
+    url = helpers.parse_url(repo_url, "issues")
 
     issues = await get_paginated_data(url, extra_params={"state": state, "sort": sort})
 
@@ -193,7 +138,7 @@ async def get_pulls(
         GitHubAPIError: If the request to GitHub failed.
     """
     pulls = []
-    url = parse_url(repo_url, "pulls")
+    url = helpers.parse_url(repo_url, "pulls")
 
     pulls = await get_paginated_data(url, extra_params={"state": state, "sort": sort})
 
@@ -218,7 +163,7 @@ async def get_contributor_history(repo_url: str) -> list[dict]:
         "X-GitHub-Api-Version": "2022-11-28",
         "Authorization": f"Bearer {GITHUB_TOKEN}",
     }
-    url = parse_url(repo_url, "stats/contributors")
+    url = helpers.parse_url(repo_url, "stats/contributors")
 
     # wait 5s, 10s, 20s, ... , 160s before giving up - total 315s for GitHub to finish
     count = 1
